// src/app/models/CadastroDomiciliar.ts
export class CadastroDomiciliar {
    codigo: number;
    cns_profissional: string;
    cbo: string;
    cnes: string;
    ine: string;
    microarea: string;
    fora_area: string;
    data_cadastro: string;
    cep: string;
    uf: string;
    municipio: string;
    bairro: string;
    tipo_logradouro: string;
    nome_logradouro: string;
    numero: number;
    sem_numero: string;
    complemento: string;
    ponto_referencia: string;
    tipo_imovel: number;
    telefone_residencial: string;
    telefone_contato: string;
    situacao_moradia: string;
    localizacao: string;
    tipo_domicilio: string;
    numero_moradores: number;
    numero_comodos: number;
    tipo_acesso_domicilio: string;
    energia_eletrica: string;
    condicao_posse_uso_terra: string;
    material_predominante_alvenaria: string;
    material_predominante_taipa: string;
    material_predominante_outros: string;
    abastecimento_agua: string;
    forma_escoamento_banheiro: string;
    agua_consumo_domicilio: string;
    destino_lixo: string;
    tem_animal: string;
    animal_gato: string;
    animal_cachorro: string;
    animal_passaro: string;
    animal_outros: string;
    animal_total: number;
    termo_recusa: string;
    termo_recusa_nome: string;
    termo_recusa_rg: string;
    nome_instituicao_permanencia: string;
    tem_outros_profissionais_saude: string;
    responsavel_tecnico_nome: string;
    responsavel_tecnico_cns: string;
    responsavel_tecnico_cargo: string;
    responsavel_tecnico_telefone: string;
    termo_recusa_instituicao_permanencia: string;
    termo_recusa_instituicao_permanencia_nome: string;
    termo_recusa_instituicao_permanencia_rg: string;
    usuario: number;
    imei: string;
    pim: string;
    versao: string;
    latitude: number;
    longitude: number;

    constructor() {
        this.codigo = 0;
        this.cns_profissional = '';
        this.cbo = '';
        this.cnes = '';
        this.ine = '';
        this.microarea = '';
        this.fora_area = 'N';
        this.data_cadastro = '';
        this.cep = '';
        this.uf = '';
        this.municipio = '';
        this.bairro = '';
        this.tipo_logradouro = '';
        this.nome_logradouro = '';
        this.numero = 0;
        this.sem_numero = 'N';
        this.complemento = '';
        this.ponto_referencia = '';
        this.tipo_imovel = 0;
        this.telefone_residencial = '';
        this.telefone_contato = '';
        this.situacao_moradia = '';
        this.localizacao = '';
        this.tipo_domicilio = '';
        this.numero_moradores = 0;
        this.numero_comodos = 0;
        this.tipo_acesso_domicilio = '';
        this.energia_eletrica = 'N';
        this.condicao_posse_uso_terra = '';
        this.material_predominante_alvenaria = '';
        this.material_predominante_taipa = '';
        this.material_predominante_outros = '';
        this.abastecimento_agua = '';
        this.forma_escoamento_banheiro = '';
        this.agua_consumo_domicilio = '';
        this.destino_lixo = '';
        this.tem_animal = 'N';
        this.animal_gato = '';
        this.animal_cachorro = '';
        this.animal_passaro = '';
        this.animal_outros = '';
        this.animal_total = 0;
        this.termo_recusa = 'N';
        this.termo_recusa_nome = '';
        this.termo_recusa_rg = '';
        this.nome_instituicao_permanencia = '';
        this.tem_outros_profissionais_saude = 'N';
        this.responsavel_tecnico_nome = '';
        this.responsavel_tecnico_cns = '';
        this.responsavel_tecnico_cargo = '';
        this.responsavel_tecnico_telefone = '';
        this.termo_recusa_instituicao_permanencia = 'N';
        this.termo_recusa_instituicao_permanencia_nome = '';
        this.termo_recusa_instituicao_permanencia_rg = '';
        this.usuario = 0;
        this.imei = '';
        this.pim = '';
        this.versao = '';
        this.latitude = 0;
        this.longitude = 0;
    }
}